JMK_140: Shifted upload code to Utils folder
